metaentry.c: Fix underallocation for xattr_lvalues.
Wrong type has been used for calculating size of requested memory.
Underallocation happened in mentries_fromfile() code path if
sizeof(ssize_t) > sizeof(int), which is true on 64-bit platforms.

Reading metadata file with extended attributes entries led to corruption
of metadata represented in memory on such platforms, so applying could
led to corruption of metadata on disk too.

Reported-by: Uros Juvan <asmpro@gmail.com>

+++ src/metaentry.c
@@ -500,7 +500,7 @@ mentries_fromfile(struct metahash **mhash, const char *path)
 		}
 
 		mentry->xattr_names   = xmalloc(mentry->xattrs * sizeof(char *));
-		mentry->xattr_lvalues = xmalloc(mentry->xattrs * sizeof(int));
+		mentry->xattr_lvalues = xmalloc(mentry->xattrs * sizeof(ssize_t));
 		mentry->xattr_values  = xmalloc(mentry->xattrs * sizeof(char *));
 
 		for (i = 0; i < mentry->xattrs; i++) {
